Break TCPSession out into its own file.
In the process, make try_yield_message() private as its not part of the interface, and remove the need for ntohs(). All hail Rob Pike. https://commandcenter.blogspot.co.uk/2012/04/byte-order-fallacy.html.
