+++ flame/trafgen.cpp
@@ -7,72 +7,12 @@
 #include <string>
 
 #include "trafgen.h"
+#include "tcpsession.h"
 
 #include <ldns/rbtree.h>
 
 #include <ldns/host2wire.h>
 #include <ldns/wire2host.h>
-
-static const size_t MIN_DNS_QUERY_SIZE = 17;
-static const size_t MAX_DNS_QUERY_SIZE = 512;
-
-class TCPSession
-{
-public:
-    using error_cb = std::function<void()>;
-    using query_cb = std::function<void(std::unique_ptr<char[]> data, size_t size)>;
-
-    bool try_yield_message()
-    {
-        uint16_t size = 0;
-
-        if (buffer.size() < sizeof(size)) {
-            return false;
-        }
-
-        memcpy(&size, buffer.data(), sizeof(size));
-        size = ntohs(size);
-
-        if (size < MIN_DNS_QUERY_SIZE || size > MAX_DNS_QUERY_SIZE) {
-            error();
-            return false;
-        }
-
-        if (buffer.size() >= sizeof(size) + size) {
-            auto data = std::make_unique<char[]>(size);
-            memcpy(data.get(), buffer.data() + sizeof(size), size);
-            buffer.erase(0, sizeof(size) + size);
-
-            query(std::move(data), size);
-            return true;
-        }
-
-        return false;
-    }
-
-    void received(const char data[], size_t len)
-    {
-        buffer.append(data, len);
-
-        while (try_yield_message()) {
-        }
-    }
-
-    void on_error(error_cb handler)
-    {
-        error = std::move(handler);
-    }
-
-    void on_query(query_cb handler)
-    {
-        query = std::move(handler);
-    }
-
-private:
-    std::string buffer;
-    error_cb error;
-    query_cb query;
-};
 
 TrafGen::TrafGen(std::shared_ptr<uvw::Loop> l,
     std::shared_ptr<Metrics> s,
